chore: refactor join, leave and play, and add stop audio commands
This commit adds the following changes:

- Refactored `join` command.
- Refactored `leave` command.
- Refactored `play` command.
- Added a `stop` command to stop the music player.

The `join` command allows the bot to join the voice channel specified by the user. If the bot is already in a voice channel, it will return without taking any action.

The `leave` command allows the bot to leave the current voice channel. If the bot is not in a voice channel, it will return without taking any action.

The `play` command allows the bot to play music in the voice channel. It takes a YouTube URL or search query as an argument and plays the corresponding audio in the voice channel. If the bot is not in a voice channel, it will return without taking any action.

The `stop` command allows the bot to stop the music player. It stops the currently playing audio and clears the queue.

These commands are implemented using the `@discordjs/voice` library for voice-related functionality and the `ytdl-core` library for YouTube audio streaming.

Note: The implementation of the `play` command includes some code duplication that can be improved in future updates.