feat: Add chalk package to dependencies
This commit adds the chalk package to the project's dependencies.
The chalk package is used for styling and coloring text in the console output.

Changes made:

- Added "chalk": "^4.1.2" to the dependencies section in both package.json and package-lock.json.

Note: This commit also includes some code changes related to debugging and logging, such as importing and using the chalk package in various files.